moved props moved tiling-container
