+++ src/app/components/diagrams/KPIs/self-consumption.component.ts
@@ -1,6 +1,6 @@
 
 import { Component, Input } from '@angular/core';
-import { Props } from '../../props';
+import { Props } from '../../../types/props';
 
 @Component({
   selector: 'app-self-consumption',
