+++ src/app/components/diagrams/energy-mix-diagram.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { Props } from '../props';
+import { Props } from '../../types/props';
 
 @Component({
   selector: 'app-energy-mix-diagram',
