+++ src/app/components/tiling-wrapper.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import {Props} from './props';
+import {Props} from '../types/props';
 
 @Component({
   selector: 'app-tiling-wrapper',
