--- src/app/views/dashboard.component.scss
@@ -1,10 +1,0 @@
-.tiling-container {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-around;
-    align-items: center;
-    width: 100%;
-    height: 100px;
-    background-color: rgba(255, 255, 255, 0);
-    border: 1px solid #000000;
-}
